contrib/git_buildbot.py: Accept a --category parameter
In order to support multiple git repositories all communicating with a
single build master, there needs to be some way to distinguish which
repository a set of changes is coming from.  This patch adds a
--category (or -c) flag to git_buildbot.py, which can be used in a
post-receive hook to set the category that changes from that
repository will be tagged with.  Schedulers may then be set up to
listen only to particular categories, and call builders based on those
categories that fetch from the correct repositories.